requirements.txt
asn1crypto==0.23.0
attrs==17.2.0
Automat==0.6.0
awscli==1.11.162
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.4.7
botocore==1.7.20
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
colorama==0.3.7
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
docutils==0.14
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
jmespath==0.9.3
lxml==4.0.0
numpy==1.13.3
pandas==0.20.3
parsel==1.2.0
psycopg2==2.7.3.1
pyasn1==0.3.6
pyasn1-modules==0.1.4
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.3.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
queuelib==1.4.2
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.11
Scrapy==1.4.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
w3lib==1.18.0
zope.interface==4.4.3
